refactored for project simplicity
