--- examples/docs_beta_snippets/docs_beta_snippets/guides/tutorials/etl_tutorial/.gitignore
@@ -1,1 +1,0 @@
-__pycache__/
