--- examples/docs_beta_snippets/docs_beta_snippets/guides/tutorials/etl_tutorial/etl_tutorial/partitions/__init__.py
@@ -1,5 +1,0 @@
-import dagster as dg
-
-monthly_partition = dg.MonthlyPartitionsDefinition(start_date="2024-01-01")
-
-product_category_partition = dg.StaticPartitionsDefinition(["Electronics", "Books", "Home and Garden", "Clothing"])
